:card_file_box: create mall models

+++ db/postgresql/migrations/2022-01-01-172016_create-nut/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users(
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     real_name VARCHAR(32) NOT NULL,
     nick_name VARCHAR(255) NOT NULL,
     email VARCHAR(255) NOT NULL,
@@ -14,7 +14,7 @@ CREATE TABLE users(
     time_zone VARCHAR(32) NOT NULL DEFAULT 'UTC',
     address VARCHAR(255),
     profile BYTEA NOT NULL,
-    sign_in_count BIGINT NOT NULL DEFAULT 0,
+    sign_in_count INTEGER NOT NULL DEFAULT 0,
     current_sign_in_at TIMESTAMP WITHOUT TIME ZONE,
     current_sign_in_ip VARCHAR(45),
     last_sign_in_at TIMESTAMP WITHOUT TIME ZONE,
@@ -22,7 +22,7 @@ CREATE TABLE users(
     confirmed_at TIMESTAMP WITHOUT TIME ZONE,
     locked_at TIMESTAMP WITHOUT TIME ZONE,
     deleted_at TIMESTAMP WITHOUT TIME ZONE,
-    version BIGINT NOT NULL DEFAULT 0,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
@@ -42,8 +42,8 @@ CREATE UNIQUE INDEX idx_users_uid ON users(uid);
 CREATE UNIQUE INDEX idx_users_provider ON users(provider_type, provider_id);
 
 CREATE TABLE "logs"(
-    id BIGSERIAL PRIMARY KEY,
-    user_id BIGINT NOT NULL,
+    id SERIAL PRIMARY KEY,
+    user_id INTEGER NOT NULL,
     LEVEL VARCHAR(16) NOT NULL DEFAULT 'INFO',
     ip VARCHAR(45) NOT NULL,
     message VARCHAR(255) NOT NULL,
@@ -51,12 +51,12 @@ CREATE TABLE "logs"(
 );
 
 CREATE TABLE groups(
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     code VARCHAR(32) NOT NULL,
     name VARCHAR(255) NOT NULL,
-    parent_id BIGINT,
-    "level" BIGINT NOT NULL DEFAULT 1,
-    version BIGINT NOT NULL DEFAULT 0,
+    parent_id INTEGER,
+    "level" INTEGER NOT NULL DEFAULT 1,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
@@ -66,21 +66,21 @@ CREATE UNIQUE INDEX idx_groups ON groups(code, "level");
 CREATE INDEX idx_groups_name ON groups(name);
 
 CREATE TABLE groups_users(
-    id BIGSERIAL PRIMARY KEY,
-    group_id BIGINT NOT NULL,
-    user_id BIGINT NOT NULL,
+    id SERIAL PRIMARY KEY,
+    group_id INTEGER NOT NULL,
+    user_id INTEGER NOT NULL,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE UNIQUE INDEX idx_groups_users ON groups_users(group_id, user_id);
 
 CREATE TABLE roles(
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     code VARCHAR(32) NOT NULL,
     name VARCHAR(255) NOT NULL,
-    parent_id BIGINT,
-    "level" BIGINT NOT NULL DEFAULT 1,
-    version BIGINT NOT NULL DEFAULT 0,
+    parent_id INTEGER,
+    "level" INTEGER NOT NULL DEFAULT 1,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
@@ -90,12 +90,12 @@ CREATE UNIQUE INDEX idx_roles ON roles(code, "level");
 CREATE INDEX idx_roles_name ON roles(name);
 
 CREATE TABLE roles_relations(
-    id BIGSERIAL PRIMARY KEY,
-    a BIGINT NOT NULL,
+    id SERIAL PRIMARY KEY,
+    a INTEGER NOT NULL,
     "constraint" VARCHAR(32) NOT NULL,
-    b BIGINT NOT NULL,
+    b INTEGER NOT NULL,
     name VARCHAR(255) NOT NULL,
-    version BIGINT NOT NULL DEFAULT 0,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
@@ -104,9 +104,9 @@ CREATE UNIQUE INDEX idx_roles_relations ON roles_relations(a, "constraint", b);
 CREATE INDEX idx_roles_relations_from ON roles_relations(a, "constraint");
 
 CREATE TABLE roles_users(
-    id BIGSERIAL PRIMARY KEY,
-    role_id BIGINT NOT NULL,
-    user_id BIGINT NOT NULL,
+    id SERIAL PRIMARY KEY,
+    role_id INTEGER NOT NULL,
+    user_id INTEGER NOT NULL,
     not_before DATE NOT NULL,
     expire_at DATE NOT NULL,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
@@ -115,9 +115,9 @@ CREATE TABLE roles_users(
 CREATE UNIQUE INDEX idx_roles_users ON roles_users(role_id, user_id);
 
 CREATE TABLE roles_groups(
-    id BIGSERIAL PRIMARY KEY,
-    role_id BIGINT NOT NULL,
-    group_id BIGINT NOT NULL,
+    id SERIAL PRIMARY KEY,
+    role_id INTEGER NOT NULL,
+    group_id INTEGER NOT NULL,
     not_before DATE NOT NULL,
     expire_at DATE NOT NULL,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
@@ -126,10 +126,10 @@ CREATE TABLE roles_groups(
 CREATE UNIQUE INDEX idx_roles_groups ON roles_groups(role_id, group_id);
 
 CREATE TABLE operations(
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     code VARCHAR(32) NOT NULL,
     name VARCHAR(255) NOT NULL,
-    version BIGINT NOT NULL DEFAULT 0,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
@@ -139,13 +139,13 @@ CREATE UNIQUE INDEX idx_operations_code ON operations(code);
 CREATE INDEX idx_operations_name ON operations(name);
 
 CREATE TABLE resources(
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     code VARCHAR(255) NOT NULL,
     TYPE VARCHAR(255) NOT NULL,
     name VARCHAR(255) NOT NULL,
-    parent_id BIGINT,
-    "level" BIGINT NOT NULL DEFAULT 1,
-    version BIGINT NOT NULL DEFAULT 0,
+    parent_id INTEGER,
+    "level" INTEGER NOT NULL DEFAULT 1,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
@@ -155,24 +155,24 @@ CREATE UNIQUE INDEX idx_resources ON resources(code, TYPE, "level");
 CREATE INDEX idx_resources_name ON resources(name);
 
 CREATE TABLE policies(
-    id BIGSERIAL PRIMARY KEY,
-    role_id BIGINT NOT NULL,
-    resource_id BIGINT NOT NULL,
-    operation_id BIGINT NOT NULL,
-    version BIGINT NOT NULL DEFAULT 0,
+    id SERIAL PRIMARY KEY,
+    role_id INTEGER NOT NULL,
+    resource_id INTEGER NOT NULL,
+    operation_id INTEGER NOT NULL,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE UNIQUE INDEX idx_policies ON policies(role_id, resource_id, operation_id);
 
 CREATE TABLE attachments(
-    id BIGSERIAL PRIMARY KEY,
-    user_id BIGINT NOT NULL,
+    id SERIAL PRIMARY KEY,
+    user_id INTEGER NOT NULL,
     title VARCHAR(255) NOT NULL,
-    size BIGINT NOT NULL,
+    size INTEGER NOT NULL,
     content_type VARCHAR(255) NOT NULL,
     url VARCHAR(255) NOT NULL,
-    version BIGINT NOT NULL DEFAULT 0,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
@@ -180,14 +180,14 @@ CREATE TABLE attachments(
 CREATE INDEX idx_attachments ON attachments(title);
 
 CREATE TABLE notifications(
-    id BIGSERIAL PRIMARY KEY,
-    user_id BIGINT NOT NULL,
+    id SERIAL PRIMARY KEY,
+    user_id INTEGER NOT NULL,
     url VARCHAR(255) NOT NULL,
     body TEXT NOT NULL,
     media_type VARCHAR(36) NOT NULL,
     "level" VARCHAR(16) NOT NULL,
     "read" BOOLEAN NOT NULL,
-    version BIGINT NOT NULL DEFAULT 0,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
@@ -195,13 +195,13 @@ CREATE TABLE notifications(
 CREATE INDEX idx_notifications ON notifications("level");
 
 CREATE TABLE tags(
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     code VARCHAR(32) NOT NULL,
     name VARCHAR(255) NOT NULL,
     font BYTEA NOT NULL,
     icon BYTEA NOT NULL,
     color BYTEA NOT NULL,
-    version BIGINT NOT NULL DEFAULT 0,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
@@ -211,11 +211,11 @@ CREATE UNIQUE INDEX idx_tags_code ON tags(code);
 CREATE INDEX idx_tags_name ON tags(name);
 
 CREATE TABLE votes(
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     resource_type VARCHAR(255) NOT NULL,
     resource_id VARCHAR(255) NOT NULL,
-    "point" BIGINT NOT NULL DEFAULT 0,
-    version BIGINT NOT NULL DEFAULT 0,
+    "point" INTEGER NOT NULL DEFAULT 0,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
@@ -225,11 +225,11 @@ CREATE UNIQUE INDEX idx_votes ON votes(resource_type, resource_id);
 CREATE INDEX idx_votes_resource_type ON votes(resource_type);
 
 CREATE TABLE view_counters(
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     resource_type VARCHAR(255) NOT NULL,
     resource_id VARCHAR(255) NOT NULL,
-    "point" BIGINT NOT NULL DEFAULT 0,
-    version BIGINT NOT NULL DEFAULT 0,
+    "point" INTEGER NOT NULL DEFAULT 0,
+    version INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
